Added include for all shared enumerations Removed all enumeration forward declarations Fixed a bunch of warnings

+++ data/cursor.h
@@ -133,7 +133,7 @@ public:
 	Icon(Cursor::CursorStyle style);
 
 	void setPosition(int x, int y);
-	void setBrightness(uint8 brightness);
+	void setBrightness(uint brightness);
 	Common::Rect draw(Graphics::Surface *surface);
 
 private:
